+++ src/Helper/EkomiHelper.php
@@ -5,6 +5,7 @@ namespace EkomiFeedback\Helper;
 use EkomiFeedback\Helper\ConfigHelper;
 use Plenty\Modules\System\Contracts\WebstoreRepositoryContract;
 use Plenty\Modules\Item\ItemImage\Contracts\ItemImageRepositoryContract;
+use Plenty\Modules\Order\Shipping\Countries\Contracts\CountryRepositoryContract;
 
 /**
  * Class EkomiHelper
@@ -17,78 +18,55 @@ class EkomiHelper {
     private $configHelper;
     private $webStoreRepo;
     private $imagesRepo;
+    private $countryRepo;
 
-    public function __construct(WebstoreRepositoryContract $webStoreRepo, ConfigHelper $configHelper, ItemImageRepositoryContract $imagesRepo) {
+    public function __construct(WebstoreRepositoryContract $webStoreRepo, ConfigHelper $configHelper, ItemImageRepositoryContract $imagesRepo, CountryRepositoryContract $countryRepo) {
         $this->configHelper = $configHelper;
         $this->webStoreRepo = $webStoreRepo;
         $this->imagesRepo = $imagesRepo;
+        $this->countryRepo = $countryRepo;
     }
 
     /**
-     * Gets the order data and prepare post variables
+     * Gets the order data and prepare post variables.
      * 
      * @param array $order Order object as array
      * 
-     * @return string The comma separated parameters
+     * @return array The comma separated parameters
      */
     function preparePostVars($order) {
-        $id = $order['id'];
-        $plentyId = $order['plentyId'];
-        $createdAt = $order['createdAt'];
+	    $id = $order['id'];
+	    $plentyId = $order['plentyId'];
+	    $fields   = array(
+		    'shop_id'            => $this->configHelper->getShopId(),
+		    'interface_password' => $this->configHelper->getShopSecret(),
+		    'mode'               => $this->configHelper->getMode(),
+		    'product_reviews'    => $this->configHelper->getProductReviews(),
+		    'plugin_name'        => 'plentymarkets',
+		    'product_identifier' => $this->configHelper->getProductIdentifier(),
+		    'exclude_products'   => $this->configHelper->getExcludeProducts(),
+	    );
 
-        $customerInfo=$order['relations'][1]['contactReceiver'];
+	    $order['senderName'] = $this->getWebStoreName($plentyId);
+        $order['senderEmail'] = '';
 
-        $apiMode = $this->getRecipientType($customerInfo['privatePhone']);
-
-        $scheduleTime = $this->toMySqlDateTime($createdAt);
-
-        $senderName = $this->getStoreName($plentyId);
-
-        if ($apiMode == 'sms' && strlen($senderName) > 11) {
-            $senderName = substr($senderName, 0, 11);
+	    foreach ($order['addresses'] as $key=>$address) {
+            $countryInfo = $this->countryRepo->getCountryById($address['countryId']);
+            $order['addresses'][$key]['countryName'] = $countryInfo->name;
+            $order['addresses'][$key]['isoCode2'] = $countryInfo->isoCode2;
+            $order['addresses'][$key]['isoCode3'] = $countryInfo->isoCode3;
         }
 
-        $fields = array(
-            'shop_id' => $this->configHelper->getShopId(),
-            'password' => $this->configHelper->getShopSecret(),
-            'recipient_type' => $apiMode,
-            'salutation' => '',
-            'first_name' => $customerInfo['firstName'],
-            'last_name' => $customerInfo['lastName'],
-            'email' => $customerInfo['email'],
-            'transaction_id' => $id,
-            'transaction_time' => $scheduleTime,
-            'telephone' => $customerInfo['privatePhone'],
-            'sender_name' => $senderName,
-            'sender_email' => ''
-        );
+        $order['orderItems']  = $this->getProductsData($order['orderItems'], $plentyId);
+	    $fields['order_data'] = $order;
 
-        $fields['client_id'] = $customerInfo['id'];
-        $fields['screen_name'] = $customerInfo['fullName'];
-
-        if ($this->configHelper->getProductReviews() == 'true') {
-            $fields['has_products'] = 1;
-            $productsData = $this->getProductsData($order['orderItems'], $plentyId);
-            $fields['products_info'] = json_encode($productsData['product_info']);
-            $fields['products_other'] = json_encode($productsData['other']);
-        }
-
-        $postVars = '';
-        $counter = 1;
-        foreach ($fields as $key => $value) {
-            if ($counter > 1)
-                $postVars .= "&";
-            $postVars .= $key . "=" . $value;
-            $counter++;
-        }
-
-        return $postVars;
+	    return $fields;
     }
 
     /**
-     * Gets item image url
+     * Gets item image url.
      * 
-     * @param type $itemId
+     * @param int $itemId
      * @return string
      */
     public function getItemImageUrl($itemId) {
@@ -139,28 +117,12 @@ class EkomiHelper {
         $products = array();
         foreach ($orderItems as $key => $product) {
             if (!empty($product['properties'])) {
-                $itemId = $product['id'];
+                $itemURLs = $this->getItemURLs($product['id'], $plentyId);
 
-                $itemURLs = $this->getItemURLs($itemId, $plentyId);
+	            $product['image_url'] = utf8_decode($itemURLs['imgUrl']);
+	            $product['canonical_url'] = utf8_decode($itemURLs['itemUrl']);
 
-                $products['product_info'][$itemId] = $product['orderItemName'];
-
-                $productOther = array();
-
-                $productOther['image_url'] = utf8_decode($itemURLs['imgUrl']);
-
-                $productOther['brand_name'] = '';
-
-                $productOther['product_ids'] = array(
-                    'gbase' => utf8_decode($itemId)
-                );
-
-                $productOther['links'] = array(
-                    array('rel' => 'canonical', 'type' => 'text/html',
-                        'href' => utf8_decode($itemURLs['itemUrl']))
-                );
-
-                $products['other'][$itemId]['product_other'] = $productOther;
+	            $products[] = $product;
             }
         }
 
@@ -168,95 +130,28 @@ class EkomiHelper {
     }
 
     /**
-     * Gets the recipient type
-     * 
-     * @param string $telephone The phone nu,ber
-     * 
-     * @return string Recipient type
-     * 
-     * @access protected
-     */
-    protected function getRecipientType($telephone) {
-
-        $reviewMod = $this->configHelper->getMod();
-        $apiMode = 'email';
-        switch ($reviewMod) {
-            case 'sms':
-                $apiMode = 'sms';
-                break;
-            case 'email':
-                $apiMode = 'email';
-                break;
-            case 'fallback':
-                if ($this->validateE164($telephone))
-                    $apiMode = 'sms';
-                else
-                    $apiMode = 'email';
-                break;
-        }
-
-        return $apiMode;
-    }
-
-    /**
-     * Validates E164 numbers
-     * 
-     * @param $phone The phone number
+     * Gets web store
      *
-     * @return bool Yes if matches False otherwise
-     * 
-     * @access protected
-     */
-    protected function validateE164($phone) {
-        $pattern = '/^\+?[1-9]\d{1,14}$/';
-
-        preg_match($pattern, $phone, $matches);
-
-        if (!empty($matches)) {
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
-     * Converts date to Mysql formate
-     * 
-     * @param string $date The date
-     * 
-     * @return string The formatted date
-     */
-    public function toMySqlDateTime($date) {
-        try {
-            return date('d-m-Y H:i:s', strtotime($date));
-        } catch (\Exception $exc) {
-            echo $exc->getTraceAsString();
-            return $date;
-        }
-    }
-
-    /**
-     * Gets store name
-     * 
      * @return string
-     * 
+     *
      * @access protected
      */
-    protected function getStoreName($plentyId) {
+    protected function getWebStoreName($plentyId) {
         $temp1 = $this->webStoreRepo->findByPlentyId($plentyId)->toArray();
         if (isset($temp1['name'])) {
             return $temp1['name'];
         }
+
         return '';
     }
 
     /**
      * Gets Store domain Url
-     * 
+     *
      * @param type $plentyId
-     * 
+     *
      * @return string
-     * 
+     *
      * @access protected
      */
     protected function getStoreDomain($plentyId) {
@@ -264,6 +159,7 @@ class EkomiHelper {
         if (isset($temp1['configuration']['domain'])) {
             return $temp1['configuration']['domain'];
         }
+
         return '';
     }